+++ extensions/gitlab/src/components/project_search.tsx
@@ -17,6 +17,7 @@ import {
   OpenProjectPipelinesPushAction,
   OpenProjectSecurityComplianceInBrowserAction,
   OpenProjectSettingsInBrowserAction,
+  OpenProjectWikiInBrowserAction,
   ProjectDefaultActions,
   ShowProjectLabels,
 } from "./project_actions";
@@ -30,7 +31,10 @@ export function ProjectListItem(props: { project: Project }): JSX.Element {
   }
   accessories.push({
     text: project.star_count.toString(),
-    icon: { source: Icon.Star, tintColor: Color.Yellow },
+    icon: {
+      source: Icon.Star,
+      tintColor: project.star_count > 0 ? Color.Yellow : null,
+    },
     tooltip: `Number of stars: ${project.star_count}`,
   });
   return (
@@ -54,6 +58,7 @@ export function ProjectListItem(props: { project: Project }): JSX.Element {
             <OpenProjectBranchesPushAction project={project} />
             <OpenProjectPipelinesPushAction project={project} />
             <OpenProjectMilestonesPushAction project={project} />
+            <OpenProjectWikiInBrowserAction project={project} />
             <ShowProjectLabels project={props.project} shortcut={{ modifiers: ["cmd"], key: "l" }} />
           </ActionPanel.Section>
           <ActionPanel.Section title="Open in Browser">