+++ extensions/music/src/util/scripts/general.ts
@@ -1,6 +1,7 @@
 import { pipe } from "fp-ts/lib/function";
 import * as R from "fp-ts/Reader";
 import * as RTE from "fp-ts/ReaderTaskEither";
+import * as TE from "fp-ts/TaskEither";
 
 import { tell } from "../apple-script";
 import { ScriptError } from "../models";
@@ -17,4 +18,11 @@ export const setShuffle: RTE.ReaderTaskEither<boolean, ScriptError, string> = pi
   R.map((shuffle) => tell("Music", `set shuffle enabled to ${shuffle.toString()}`))
 );
 
-export const getLibraryName = tell("Music", "get name of source 1");
+export const getLibraryName = pipe(
+  tell("Music", "get name of source 1"),
+  TE.orElse((err) => {
+    console.error(err);
+    // fallback to "Library"
+    return TE.right("Library");
+  })
+);