+++ extensions/lastpass/CHANGELOG.md
@@ -1,5 +1,9 @@
 # LastPass Changelog
 
+## [Enhancement] - 2023-03-15
+
+- Add ability to search by ID
+
 ## [Enhancement] - 2023-03-13
 
 - Optimized the code